speech_requirements.txt
# Speech dependencies for natural-sounding voice and speech recognition
elevenlabs>=0.2.24
pygame>=2.5.2
aiohttp>=3.8.0

# Speech-to-text dependencies
SpeechRecognition>=3.10.0
pyaudio>=0.2.11

# Existing dependencies (if not already installed)
browser-use
openai
python-dotenv
pydantic